// src/types/api.ts
export interface Credential {
  name: string
  exists: boolean
  type: "text" | "file"
  fileName?: string
  modifiedAt: string
}